backend/requirements.txt
flask
mysql-connector-python
pyjwt
python-dotenv
flask_cors
moviepy
opencv-python
gunicorn